\name{autophagy}
\alias{autophagy}
\docType{data}
\title{
Data set for autophgay study.
}
\description{
Data set used to fit a Poisson regression from an autophgay study.
}
\usage{data(autophagy)}
\format{
  The format is:
 chr "autophagy"
}
\examples{
data(autophagy)
## maybe str(autophagy) ; plot(autophagy) ...
}
\keyword{datasets}
