\name{colocolization}
\alias{colocolization}
\docType{data}
\title{
Data set for autophgay study.
}
\description{
Data set used to fit a customized funcion from an autophgay study.
}
\usage{data(colocolization)}
\format{
  The format is:
 chr "colocolization"
}
\examples{
data(colocolization)
## maybe str(colocolization) ; plot(colocolization) ...
}
\keyword{datasets}
