\name{iScreenInBatchBoxplot}
\alias{iScreenInBatchBoxplot}
\title{Plotting iScreenInBatch boxplot}
\usage{
  iScreenInBatchBoxplot(object, ...)
}
\arguments{
  \item{object}{A iScreenInBatch object.}

  \item{...}{Arguments to be passed to methods. See
  \code{\link{boxplot}} and \code{\link{par}}}
}
\description{
  Function for plotting object returned by iScreenInBatch.
  For more details about the graphical parameter arguments,
  see \code{\link{par}}.
}

