\name{iWellPlot}
\alias{iWellPlot}
\title{Plotting iWell}
\usage{
  iWellPlot(object, xlab = "X", ylab = "Y", ...)
}
\arguments{
  \item{object}{A iScreen object.}

  \item{xlab}{Default is "X".}

  \item{ylab}{Default is "Y".}

  \item{...}{Arguments to be passed to methods. See
  \code{\link{plot}} and \code{\link{par}}}
}
\description{
  Function for plotting object returned by iPlate. For more
  details about the graphical parameter arguments, see
  \code{\link{par}}.
}

